fix(money): store and compute money as Decimal
Per-token prices and wallet amounts were double precision, so ten 0.1 USD
payments summed to 0.9999999999999999 and the MCP payment transport refused
the third dime of a 0.3 budget. Migrate model_specs prices to NUMERIC(20,12)
and wallet budget/amount to NUMERIC(18,6) (text cast, reversible), and carry
Decimal through the ORM, wallet ledger sums, payment clients, execution
payment path, governance budget gates and the SDK cost computation. Wallet and
model price API fields serialize as strings, per the Money convention; the
OpenAPI spec and generated clients are regenerated.

Real-money defaults removed: MPP session_budget_usd (10.0 in the API schema,
MPP client and unified client, and the webapp form), the unused
TASK__DEFAULT_TASK_BUDGET_USD=1.0 setting, the 5.0 service budget prefilled
in the wallet forms, and the 0 server default on model price columns that
turned an unpriced model into a free one.

AuditObserver is deleted: it had no sink anywhere, only saw calls that every
gate had already allowed, and would have written an "allow" row per LLM and
tool call rather than a governance decision.

Fixes #504

+++ agentarea-platform/libs/llm/agentarea_llm/domain/models.py
@@ -1,9 +1,10 @@
+from decimal import Decimal
 from importlib import import_module
 from uuid import UUID
 
 from agentarea_common.base.models import BaseModel, WorkspaceScopedMixin
 from agentarea_common.constants import MANAGED_BY_PLATFORM, PLATFORM_WORKSPACE_ID
-from sqlalchemy import Boolean, Float, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import Boolean, ForeignKey, Integer, Numeric, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -142,8 +143,8 @@ class ModelSpec(BaseModel, WorkspaceScopedMixin):
     description: Mapped[str | None] = mapped_column(String, nullable=True)
     context_window: Mapped[int] = mapped_column(Integer, nullable=False)
     max_output_tokens: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    input_cost_per_token: Mapped[float | None] = mapped_column(Float, nullable=True)
-    output_cost_per_token: Mapped[float | None] = mapped_column(Float, nullable=True)
+    input_cost_per_token: Mapped[Decimal | None] = mapped_column(Numeric(20, 12), nullable=True)
+    output_cost_per_token: Mapped[Decimal | None] = mapped_column(Numeric(20, 12), nullable=True)
     supports_function_calling: Mapped[bool | None] = mapped_column(
         Boolean, nullable=True, default=False
     )
